requirements.txt
flask==2.3.3
flask-cors==4.0.0
google-generativeai==0.3.2
python-dotenv==1.0.0
flask-sqlalchemy==3.0.5
requests==2.31.0
